fix: major refinements to code

+++ ratatui/rag_config.rb
@@ -10,7 +10,7 @@ module RAGComponents
     @rag_editing = false
     @rag_setup_error = nil
     @rag_process_error = nil
-    @rag_process_error_msg = ''
+    @rag_process_error_msg = nil
     @rag_process_running = false
     @rag_process_success = false
     @rag_frame_count = 0
@@ -22,35 +22,44 @@ module RAGComponents
     case event 
     in {type: :key, code: 'enter'}
       return true if @rag_process_running
+      @rag_process_error_msg = nil
       @rag_process_success = false
-      @rag_process_error = nil 
+      @rag_process_error = false
       @rag_process_running = true
       @rag_frame_count = 0
 
-      begin
-        case @rag_focused
-        when 0
-          api_tokens = File.read(@rag_components['api_token_path']).split('\n').map(&:strip)
-          err = Setup.init_rag_db!(@rag_components, api_tokens)
-          @rag_process_error = err unless err.nil?
-        when 1
-          raise "DB doesn't currently exist." if !File.exist?(@rag_components['api_token_path'])
-          FileUtils.rm_rf(Setup::RAG_DB_FILEPATH)
-        else  
-          raise 'unknown index found.'
-        end
-        @rag_process_success = true
-      rescue StandardError => e 
-        @rag_process_error_msg = e.message
-        @rag_process_error = true
-        @rag_process_success = false
-      ensure 
-        @rag_process_running = false
+      Thread.new do 
+        begin
+          case @rag_focused
+          when 0
+            if File.exist?(@rag_components.fetch('rag_db'))
+              @rag_process_error_msg = "RAG DB already exists."
+              @rag_process_error = true;
+            else 
+              api_tokens = File.readlines(@rag_components['api_token_path']).map(&:strip)
+              err = Setup.init_rag_db!(@rag_components, api_tokens)
+              @rag_process_error_msg = err unless err.nil?
+            end 
+          when 1
+            @rag_process_error_msg = "DB doesn't currently exist." if !File.exist?(@rag_components['rag_db'])
+            FileUtils.rm_rf(Setup::RAG_DB_FILEPATH) if @rag_process_error_msg.nil?
+          else  
+            raise 'unknown index found.'
+          end
+          @rag_process_error = !@rag_process_error_msg.nil?
+          @rag_process_success = !@rag_process_error
+        rescue StandardError => e 
+          @rag_process_error_msg = e
+          @rag_process_error = true
+          @rag_process_success = false
+        ensure 
+          @rag_process_running = false
+        end 
       end 
       true
     in {type: :key, code: 'esc'}
       @rag_editing = false 
-      @rag_process_error = nil 
+      @rag_process_error = false 
       @rag_process_running = false 
       @rag_process_success = false
       true 
@@ -76,7 +85,7 @@ module RAGComponents
     )
 
     # --- Render the state variables --- 
-    current_state = ['Current variables of interest: '] + @rag_components.map do |k, v| 
+    current_state = ['Current variables of interest (change if wrong): '] + @rag_components.map do |k, v| 
       ['', tui.text(:span, content: "#{k}: #{v}", style: tui.style(fg: :dark_gray))]
     end
     # --- END ---
@@ -106,14 +115,13 @@ module RAGComponents
     # --- END ---
 
     # --- Render the status area ---
-    # content, style = nil, nil
     if @rag_process_running
       @rag_frame_count += 1
       content = "Running process now"
       content += '.' * ((@rag_frame_count % NUM_DOTS) + 1)
       style = tui.style(fg: :yellow)
     elsif @rag_process_error
-      content = "An error happened with the process: #{@rag_process_error_msg}"
+      content = @rag_process_error_msg
       style = tui.style(fg: :red)
     elsif @rag_process_success
       content = 'Process successful!'
@@ -126,7 +134,7 @@ module RAGComponents
         text: tui.text(:span, content: content, style: style),
         alignment: :left, wrap: true, 
         block: tui.block(title: ' Status ', borders: :all, padding: 1)
-      ), status_area)
+    ), status_area)
     # --- END ---
     
   end
